refactor(scope): move unstringification of BigInts to Library from client
The CLI uses the `unstringifyBigInts` utility on Objects and then passes
it to the library functions. When working on a webclient or using the library
without the cli this leads to production of errors, one example is
generating callData for Solidity Verifier, since unstringification of
BigBigInts haven’t happened, the callData generated is corrupt, anyone
who wants to fix this would have to read the cli to figure out what is
going wrong.

Two files: `plonk_verify.js` & `groth16_verify.js` are doing this
unstringification in themselves already, and the CLI is doing it again
making the unstringification reduntant.

This PR moves the unstringification of BigInts to library files, which saves the
effort for anyone consuming the library to dig deep in the CLI and then
figure our what is going wrong, it also solves the reduntancy mentioned
in the section above.

The PR also includes the updated builds and has also passed all
described library tests.